# crates/web-host/openapi.yaml
...
openapi: 3.1.0
info:
  title: mooR Web Host API
  version: "1.0"
  description: |
    HTTP, WebSocket, and SSE API for the mooR virtual-world server.

    **Content negotiation** — Most `/v1/*` endpoints support two response
    formats selected via the `Accept` header:

    | Accept value                    | Response format   |
    |---------------------------------|-------------------|
    | `application/x-flatbuffers`     | FlatBuffers binary (default) |
    | `application/json`              | JSON              |

    If neither is acceptable, the server returns `406 Not Acceptable`.

    **Authentication** — Authenticated endpoints require an
    `X-Moor-Auth-Token` header containing a PASETO v4.public token obtained
    from `/auth/connect` or `/auth/create`.  Some endpoints also require
    `X-Moor-Client-Token` and `X-Moor-Client-Id` headers for session
    continuity.

    **FlatBuffers note** — FlatBuffer request/response bodies are opaque
    binary blobs whose schemas are defined in the `moor-schema` crate.
    This spec documents their logical content in prose; the wire format is
    governed by the `.fbs` files, not JSON Schema.

  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /auth/connect:
    post:
      operationId: authConnect
      tags: [Auth]
      summary: Authenticate an existing player
      description: |
        Runs the `connect` login command against the daemon.  On success the
        response body is a FlatBuffer `LoginResult` and the three session
        headers are set.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Login successful
          headers:
            X-Moor-Auth-Token:
              schema: { type: string }
              description: PASETO auth token for subsequent requests
            X-Moor-Client-Token:
              schema: { type: string }
              description: Session token for this connection
            X-Moor-Client-Id:
              schema: { type: string, format: uuid }
              description: Unique ID for this client session
          content:
            application/x-flatbuffers:
              schema: { type: string, format: binary }
              # Logical content: LoginResult { success, auth_token, player, player_flags }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Login rejected by daemon
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          description: Malformed response from daemon
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/create:
    post:
      operationId: authCreate
      tags: [Auth]
      summary: Create a new player account
      description: |
        Runs the `create` login command against the daemon.  Identical to
        `/auth/connect` except the daemon verb is `:do_login_command("create", …)`.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Account created and logged in
          headers:
            X-Moor-Auth-Token:
              schema: { type: string }
            X-Moor-Client-Token:
              schema: { type: string }
            X-Moor-Client-Id:
              schema: { type: string, format: uuid }
          content:
            application/x-flatbuffers:
              schema: { type: string, format: binary }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Account creation rejected by daemon
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          description: Malformed response from daemon
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/validate:
    get:
      operationId: authValidate
      tags: [Auth]
      summary: Validate an auth token
      description: |
        Round-trips to the daemon to verify the token is still valid.
        An ephemeral attach/detach is performed; the connection is cleaned
        up automatically.
      security:
        - MoorAuthToken: []
      responses:
        "200":
          description: Token is valid (empty body)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/logout:
    post:
      operationId: authLogout
      tags: [Auth]
      summary: Explicitly log out a player session
      description: |
        Sends a detach message to the daemon with `disconnected=true`,
        triggering `user_disconnected`.  Requires both the auth token and
        client credentials.
      security:
        - MoorAuthToken: []
      parameters:
        - $ref: "#/components/parameters/ClientToken"
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: Logout successful (empty body)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── OAuth2 ────────────────────────────────────────────────────────────
  /v1/oauth2/config:
    get:
      operationId: oauth2Config
      tags: [OAuth2]
      summary: Get OAuth2 configuration
      description: |
        Returns whether OAuth2 is enabled and the list of configured
        providers.  Route is only present when OAuth2 is enabled.
      responses:
        "200":
          description: OAuth2 configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2ConfigResponse"
        "404":
          description: OAuth2 not enabled (route absent)
  /auth/oauth2/{provider}/authorize:
    get:
      operationId: oauth2Authorize
      tags: [OAuth2]
      summary: Start a cookie-bound OAuth2 authorization flow
      description: |
        Generates a provider authorization URL and a CSRF state token.
        Sets a `moor_oauth_nonce` cookie for cookie-bound redemption.
      parameters:
        - $ref: "#/components/parameters/OAuthProvider"
      responses:
        "200":
          description: Authorization URL generated
          headers:
            Set-Cookie:
              schema: { type: string }
              description: "`moor_oauth_nonce` HttpOnly cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUrlResponse"
        "400":
          description: Unknown provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: OAuth2 not enabled
  /auth/oauth2/{provider}/app/start:
    post:
      operationId: oauth2AppStart
      tags: [OAuth2]
      summary: Start a proof-bound OAuth2 authorization flow
      description: |
        Starts OAuth2 for clients that do not rely on browser cookies.
        Stores redirect target + PKCE challenge server-side and returns an
        authorization URL. Callback redirects to the app redirect URI with a
        one-time `handoff_code`.
      parameters:
        - $ref: "#/components/parameters/OAuthProvider"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppStartRequest"
      responses:
        "200":
          description: Authorization URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppAuthUrlResponse"
        "400":
          description: Invalid provider, redirect URI, or PKCE input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: OAuth2 not enabled
  /auth/oauth2/{provider}/callback:
    get:
      operationId: oauth2Callback
      tags: [OAuth2]
      summary: OAuth2 provider callback
      description: |
        Called by the OAuth2 provider after user consent.  Validates the
        CSRF state, exchanges the authorization code for user info, and
        redirects according to flow binding:

        - **Cookie-bound existing user** → `/#auth_code={code}`
        - **Cookie-bound new user** → `/#oauth2_code={code}&oauth2_display={json}`
        - **Proof-bound** → `{redirect_uri}?handoff_code={code}`

        The code is a server-side one-time token redeemable via
        `/auth/oauth2/exchange` (cookie) or `/auth/oauth2/app/exchange` (proof).
      parameters:
        - $ref: "#/components/parameters/OAuthProvider"
        - name: code
          in: query
          required: true
          schema: { type: string }
          description: Authorization code from the OAuth2 provider
        - name: state
          in: query
          required: true
          schema: { type: string }
          description: CSRF state token
      responses:
        "302":
          description: Redirect to client with one-time code in fragment
        "404":
          description: OAuth2 not enabled (route absent)
  /auth/oauth2/exchange:
    post:
      operationId: oauth2Exchange
      tags: [OAuth2]
      summary: Exchange a cookie-bound one-time code for tokens or identity
      description: |
        Redeems a server-side one-time code produced by the callback.
        Returns either an `auth_session` (existing user) or an `identity`
        (new user needing account choice).  Requires the `moor_oauth_nonce`
        cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeExchangeRequest"
      responses:
        "200":
          description: Code exchanged
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/OAuth2AuthSessionResponse"
                  - $ref: "#/components/schemas/OAuth2IdentityResponse"
                discriminator:
                  propertyName: type
        "401":
          description: Invalid/expired code or missing nonce cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: OAuth2 not enabled (route absent)
  /auth/oauth2/app/exchange:
    post:
      operationId: oauth2AppExchange
      tags: [OAuth2]
      summary: Exchange a proof-bound handoff code
      description: |
        Redeems a one-time proof-bound handoff code using a PKCE verifier.
        Returns an auth session for existing users, or identity details plus
        an `identity_code` for account-choice submission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppExchangeRequest"
      responses:
        "200":
          description: Code exchanged
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/OAuth2AuthSessionResponse"
                  - $ref: "#/components/schemas/OAuth2AppIdentityResponse"
                discriminator:
                  propertyName: type
        "401":
          description: Invalid/expired code or invalid verifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: OAuth2 not enabled (route absent)
  /auth/oauth2/account:
    post:
      operationId: oauth2AccountChoice
      tags: [OAuth2]
      summary: Create or link an account after OAuth2
      description: |
        Submits the user's account-choice decision (create new player or
        link to existing player).  The `oauth2_code` must resolve to a
        verified `Identity` on the server side.  Requires `moor_oauth_nonce`
        cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountChoiceRequest"
      responses:
        "200":
          description: Account created/linked, session established
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2LoginResponse"
        "400":
          description: Invalid mode or code type mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication failed or invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2LoginResponse"
        "404":
          description: OAuth2 not enabled
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/oauth2/app/account:
    post:
      operationId: oauth2AppAccountChoice
      tags: [OAuth2]
      summary: Create or link account using proof-bound identity code
      description: |
        Submits account-choice for proof-bound OAuth2 flows.
        Requires `identity_code` and PKCE `code_verifier`; no nonce cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppAccountChoiceRequest"
      responses:
        "200":
          description: Account created/linked, session established
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2LoginResponse"
        "400":
          description: Invalid mode or code type mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication failed or invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2LoginResponse"
        "404":
          description: OAuth2 not enabled
        "500":
          $ref: "#/components/responses/InternalError"

  # ── WebSocket ─────────────────────────────────────────────────────────
  /ws/attach/connect:
    get:
      operationId: wsAttachConnect
      tags: [WebSocket]
      summary: WebSocket upgrade — attach with existing player
      description: |
        Upgrades to a WebSocket connection for an already-authenticated
        player (the `connect` login path).

        Auth credentials can be passed as HTTP headers or via
        `Sec-WebSocket-Protocol` subprotocols:

        | Subprotocol prefix | Equivalent header       |
        |--------------------|-------------------------|
        | `paseto.{token}`   | `X-Moor-Auth-Token`    |
        | `client_id.{uuid}` | `X-Moor-Client-Id`     |
        | `client_token.{t}` | `X-Moor-Client-Token`  |
        | `initial_attach.{bool}` | (session hint)    |

        The server selects `moor` as the WebSocket subprotocol.

        **Note:** Full WebSocket message framing is FlatBuffer-based and
        is outside the scope of this HTTP spec.
      security:
        - MoorAuthToken: []
      responses:
        "101":
          description: WebSocket upgrade successful
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /ws/attach/create:
    get:
      operationId: wsAttachCreate
      tags: [WebSocket]
      summary: WebSocket upgrade — attach with new player
      description: |
        Same as `/ws/attach/connect` but used when the player was created
        via `/auth/create`.  See that endpoint for auth details.
      security:
        - MoorAuthToken: []
      responses:
        "101":
          description: WebSocket upgrade successful
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Commands and eval ─────────────────────────────────────────────────
  /v1/command:
    post:
      operationId: executeCommand
      tags: [Commands]
      summary: Execute a MOO command and return its result
      description: |
        The daemon parses and executes one MOO command as the authenticated player.
        The request does not create a persistent connection. Input requests fail.
        Output from forked tasks is not included.

        The daemon processes the command in this order:

        1. If `#0:do_command` exists, the daemon calls it first.
        2. If `#0:do_command` returns a true value, the task succeeds with that value.
        3. If the verb is absent or returns false, the daemon uses the normal command parser.
        4. If the parser rejects the input, the outcome is `InvocationError` with
           `CommandExecutionError(CouldNotParseCommand)`.
        5. If no command verb matches and the player has a location, the daemon calls `:huh` there.
        6. If the player has no location or `:huh` does not exist, the outcome is `InvocationError` with
           `CommandExecutionError(NoCommandMatch)`.

        Any invoked verb can return a value or raise an error. This includes `#0:do_command`, the
        selected command verb, and `:huh`. The result becomes the invocation outcome. The protocol
        does not create an "I do not understand" message. If `:huh` creates this message, it is
        ordinary narrative output.

        Permission and database errors use the corresponding `CommandExecutionError` variant.
        Deadline, cancellation, and verb execution errors use their normal scheduler-error
        variants. All these task errors appear inside `InvocationError`.

        A valid HTTP request receives status 200 after the task finishes. HTTP status 200 does not
        mean that the MOO task succeeded. The `InvocationResponse.outcome` field identifies
        `InvocationSuccess` or `InvocationError`. Task errors remain inside `InvocationError`. The
        `output` field contains committed root-task output for both outcomes. It can contain output
        committed before a later task error.

        HTTP status 400 means that the HTTP request is invalid. This includes an invalid timeout or
        a body that is not valid UTF-8. A MOO parse error is a completed task response with HTTP
        status 200.
      security:
        - MoorAuthToken: []
      parameters:
        - $ref: "#/components/parameters/CaptureTimeoutMs"
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string }
            example: "look"
      responses:
        "200":
          description: |
            Completed command invocation. `InvocationResponse.outcome` identifies task success or
            error. The HTTP status remains 200 for a completed `InvocationError` response.
          content:
            application/x-flatbuffers:
              schema: { type: string, format: binary }
              # Logical: InvocationResponse { InvocationSuccess | InvocationError, output }
            application/json:
              schema:
                type: object
        "400":
          description: Invalid timeout, or the request body is not valid UTF-8
        "401":
          $ref: "#/components/responses/Unauthorized"
        "406":
          $ref: "#/components/responses/NotAcceptable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/eval:
    post:
      operationId: eval
      tags: [Eval]
      summary: Evaluate a MOO expression
      description: |
        The daemon evaluates a MOO expression as the authenticated player. The request creates no
        daemon connection. The response contains the result or error and committed root-task
        output.
      security:
        - MoorAuthToken: []
      parameters:
        - $ref: "#/components/parameters/CaptureTimeoutMs"
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string }
            example: "player:name()"
      responses:
        "200":
          description: Evaluation result or task error with committed narrative output
          content:
            application/x-flatbuffers:
              schema: { type: string, format: binary }
              # Logical: InvocationResponse { InvocationSuccess | InvocationError, output }
            application/json:
              schema:
                type: object
        "400":
          description: Invalid timeout
        "401":
          $ref: "#/components/responses/Unauthorized"
        "406":
          $ref: "#/components/responses/NotAcceptable"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ── Verbs ─────────────────────────────────────────────────────────────
...
  /v1/objects:
    get:
      operationId: listObjects
      tags: [Objects]
      summary: List all objects visible to the player
      security:
        - MoorAuthToken: []
      responses:
        "200":
          description: Object listing
          content:
            application/x-flatbuffers:
              schema: { type: string, format: binary }
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "406":
          $ref: "#/components/responses/NotAcceptable"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    MoorAuthToken:
      type: apiKey
      in: header
      name: X-Moor-Auth-Token
      description: PASETO v4.public token from `/auth/connect` or `/auth/create`
...
  schemas:
    AuthRequest:
      type: object
      properties:
        player:
          type: string
          description: Player name
        password:
          type: string
          description: Player password
      required: [player, password]
...
  responses:
    Unauthorized:
      description: Missing or invalid auth token
...
    InternalError:
      description: Internal server error
    ServiceUnavailable:
      description: Daemon is unreachable